Extract bottom tab screen
